+++ packages/google_mobile_ads/example/lib/main.dart
@@ -14,13 +14,18 @@
 
 // ignore_for_file: public_member_api_docs
 
-import 'dart:async';
+import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 
-import 'constants.dart';
 import 'reusable_inline_example.dart';
+
+void main() {
+  WidgetsFlutterBinding.ensureInitialized();
+  MobileAds.instance.initialize();
+  runApp(MyApp());
+}
 
 // You can also test with your own ad unit IDs by registering your device as a
 // test device. Check the logs for your device's ID value.
@@ -45,23 +50,17 @@ class _MyAppState extends State<MyApp> {
   RewardedAd? _rewardedAd;
   int _numRewardedLoadAttempts = 0;
 
+  BannerAd? _anchoredBanner;
+  bool _loadingAnchoredBanner = false;
+
   @override
   void initState() {
     super.initState();
-    MobileAds.instance.initialize().then((InitializationStatus status) {
-      print('Initialization done: ${status.adapterStatuses}');
-      MobileAds.instance
-          .updateRequestConfiguration(RequestConfiguration(
-              tagForChildDirectedTreatment:
-                  TagForChildDirectedTreatment.unspecified))
-          .then((void value) {
-        createInterstitialAd();
-        createRewardedAd();
-      });
-    });
+    _createInterstitialAd();
+    _createRewardedAd();
   }
 
-  void createInterstitialAd() {
+  void _createInterstitialAd() {
     InterstitialAd.load(
         adUnitId: InterstitialAd.testAdUnitId,
         request: request,
@@ -76,13 +75,13 @@ class _MyAppState extends State<MyApp> {
             _numInterstitialLoadAttempts += 1;
             _interstitialAd = null;
             if (_numInterstitialLoadAttempts <= maxFailedLoadAttempts) {
-              createInterstitialAd();
+              _createInterstitialAd();
             }
           },
         ));
   }
 
-  void showInterstitialAd() {
+  void _showInterstitialAd() {
     if (_interstitialAd == null) {
       print('Warning: attempt to show interstitial before loaded.');
       return;
@@ -93,19 +92,19 @@ class _MyAppState extends State<MyApp> {
       onAdDismissedFullScreenContent: (InterstitialAd ad) {
         print('$ad onAdDismissedFullScreenContent.');
         ad.dispose();
-        createInterstitialAd();
+        _createInterstitialAd();
       },
       onAdFailedToShowFullScreenContent: (InterstitialAd ad, AdError error) {
         print('$ad onAdFailedToShowFullScreenContent: $error');
         ad.dispose();
-        createInterstitialAd();
+        _createInterstitialAd();
       },
     );
     _interstitialAd!.show();
     _interstitialAd = null;
   }
 
-  void createRewardedAd() {
+  void _createRewardedAd() {
     RewardedAd.load(
         adUnitId: RewardedAd.testAdUnitId,
         request: request,
@@ -120,13 +119,13 @@ class _MyAppState extends State<MyApp> {
             _rewardedAd = null;
             _numRewardedLoadAttempts += 1;
             if (_numRewardedLoadAttempts <= maxFailedLoadAttempts) {
-              createRewardedAd();
+              _createRewardedAd();
             }
           },
         ));
   }
 
-  void showRewardedAd() {
+  void _showRewardedAd() {
     if (_rewardedAd == null) {
       print('Warning: attempt to show rewarded before loaded.');
       return;
@@ -137,12 +136,12 @@ class _MyAppState extends State<MyApp> {
       onAdDismissedFullScreenContent: (RewardedAd ad) {
         print('$ad onAdDismissedFullScreenContent.');
         ad.dispose();
-        createRewardedAd();
+        _createRewardedAd();
       },
       onAdFailedToShowFullScreenContent: (RewardedAd ad, AdError error) {
         print('$ad onAdFailedToShowFullScreenContent: $error');
         ad.dispose();
-        createRewardedAd();
+        _createRewardedAd();
       },
     );
 
@@ -152,18 +151,59 @@ class _MyAppState extends State<MyApp> {
     _rewardedAd = null;
   }
 
+  Future<void> _createAnchoredBanner(BuildContext context) async {
+    final AnchoredAdaptiveBannerAdSize? size =
+        await AdSize.getAnchoredAdaptiveBannerAdSize(
+      Orientation.portrait,
+      MediaQuery.of(context).size.width.truncate(),
+    );
+
+    if (size == null) {
+      print('Unable to get height of anchored banner.');
+      return;
+    }
+
+    final BannerAd banner = BannerAd(
+      size: size,
+      request: request,
+      adUnitId: Platform.isAndroid
+          ? 'ca-app-pub-3940256099942544/6300978111'
+          : 'ca-app-pub-3940256099942544/2934735716',
+      listener: BannerAdListener(
+        onAdLoaded: (Ad ad) {
+          print('$BannerAd loaded.');
+          setState(() {
+            _anchoredBanner = ad as BannerAd?;
+          });
+        },
+        onAdFailedToLoad: (Ad ad, LoadAdError error) {
+          print('$BannerAd failedToLoad: $error');
+          ad.dispose();
+        },
+        onAdOpened: (Ad ad) => print('$BannerAd onAdOpened.'),
+        onAdClosed: (Ad ad) => print('$BannerAd onAdClosed.'),
+      ),
+    );
+    return banner.load();
+  }
+
   @override
   void dispose() {
+    super.dispose();
     _interstitialAd?.dispose();
     _rewardedAd?.dispose();
-    super.dispose();
+    _anchoredBanner?.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Builder(
-        builder: (BuildContext context) => Scaffold(
+      home: Builder(builder: (BuildContext context) {
+        if (!_loadingAnchoredBanner) {
+          _loadingAnchoredBanner = true;
+          _createAnchoredBanner(context);
+        }
+        return Scaffold(
           appBar: AppBar(
             title: const Text('AdMob Plugin example app'),
             actions: <Widget>[
@@ -171,18 +211,10 @@ class _MyAppState extends State<MyApp> {
                 onSelected: (String result) {
                   switch (result) {
                     case 'InterstitialAd':
-                      showInterstitialAd();
+                      _showInterstitialAd();
                       break;
                     case 'RewardedAd':
-                      showRewardedAd();
-                      break;
-                    case 'ReusableInlineExample':
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute<void>(
-                            builder: (BuildContext context) =>
-                                ReusableInlineExample()),
-                      );
+                      _showRewardedAd();
                       break;
                     default:
                       throw AssertionError('unexpected button: $result');
@@ -197,274 +229,27 @@ class _MyAppState extends State<MyApp> {
                     value: '$RewardedAd',
                     child: Text('$RewardedAd'),
                   ),
-                  PopupMenuItem<String>(
-                    value: 'ReusableInlineExample',
-                    child: Text('Reusable Inline Ads Object Example'),
-                  ),
                 ],
               ),
             ],
           ),
-          body: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16.0),
-            child: ListView.separated(
-              cacheExtent: 500,
-              itemCount: 100,
-              separatorBuilder: (BuildContext context, int index) {
-                return Container(
-                  height: 40,
-                );
-              },
-              itemBuilder: (BuildContext context, int index) {
-                if (index % 2 == 0) {
-                  return Text(
-                    Constants.placeholderText,
-                    style: TextStyle(fontSize: 24),
-                  );
-                }
-
-                final int adIndex = (index / 2).floor();
-                Widget adWidget;
-                if (adIndex % 3 == 0) {
-                  adWidget = BannerAdWidget(AdSize.banner);
-                } else if (adIndex % 3 == 1) {
-                  adWidget = AdManagerBannerAdWidget(AdSize.largeBanner);
-                } else {
-                  adWidget = NativeAdWidget();
-                }
-                return Center(child: adWidget);
-              },
+          body: SafeArea(
+            child: Stack(
+              alignment: AlignmentDirectional.bottomCenter,
+              children: <Widget>[
+                ReusableInlineExample(),
+                if (_anchoredBanner != null)
+                  Container(
+                    color: Colors.green,
+                    width: _anchoredBanner!.size.width.toDouble(),
+                    height: _anchoredBanner!.size.height.toDouble(),
+                    child: AdWidget(ad: _anchoredBanner!),
+                  ),
+              ],
             ),
           ),
-        ),
-      ),
-    );
-  }
-}
-
-class BannerAdWidget extends StatefulWidget {
-  BannerAdWidget(this.size);
-
-  final AdSize size;
-
-  @override
-  State<StatefulWidget> createState() => BannerAdState();
-}
-
-class BannerAdState extends State<BannerAdWidget> {
-  late BannerAd _bannerAd;
-  final Completer<BannerAd> bannerCompleter = Completer<BannerAd>();
-
-  @override
-  void initState() {
-    super.initState();
-    _bannerAd = BannerAd(
-      adUnitId: BannerAd.testAdUnitId,
-      request: AdRequest(),
-      size: widget.size,
-      listener: BannerAdListener(
-        onAdLoaded: (Ad ad) {
-          print('$BannerAd loaded.');
-          bannerCompleter.complete(ad as BannerAd);
-        },
-        onAdFailedToLoad: (Ad ad, LoadAdError error) {
-          ad.dispose();
-          print('$BannerAd failedToLoad: $error');
-          bannerCompleter.completeError(error);
-        },
-        onAdOpened: (Ad ad) => print('$BannerAd onAdOpened.'),
-        onAdClosed: (Ad ad) => print('$BannerAd onAdClosed.'),
-      ),
-    );
-    Future<void>.delayed(Duration(seconds: 1), () => _bannerAd.load());
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _bannerAd.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<BannerAd>(
-      future: bannerCompleter.future,
-      builder: (BuildContext context, AsyncSnapshot<BannerAd> snapshot) {
-        Widget child;
-
-        switch (snapshot.connectionState) {
-          case ConnectionState.none:
-          case ConnectionState.waiting:
-          case ConnectionState.active:
-            child = Container();
-            break;
-          case ConnectionState.done:
-            if (snapshot.hasData) {
-              child = AdWidget(ad: _bannerAd);
-            } else {
-              child = Text('Error loading $BannerAd');
-            }
-        }
-
-        return Container(
-          width: _bannerAd.size.width.toDouble(),
-          height: _bannerAd.size.height.toDouble(),
-          color: Colors.blueGrey,
-          child: child,
         );
-      },
+      }),
     );
   }
-}
-
-class AdManagerBannerAdWidget extends StatefulWidget {
-  AdManagerBannerAdWidget(this.size);
-
-  final AdSize size;
-
-  @override
-  State<StatefulWidget> createState() => AdManagerBannerAdState();
-}
-
-class AdManagerBannerAdState extends State<AdManagerBannerAdWidget> {
-  late AdManagerBannerAd _bannerAd;
-  final Completer<AdManagerBannerAd> bannerCompleter =
-      Completer<AdManagerBannerAd>();
-
-  @override
-  void initState() {
-    super.initState();
-    _bannerAd = AdManagerBannerAd(
-      adUnitId: '/6499/example/banner',
-      request: AdManagerAdRequest(nonPersonalizedAds: true),
-      sizes: <AdSize>[widget.size],
-      listener: AdManagerBannerAdListener(
-        onAdLoaded: (Ad ad) {
-          print('$AdManagerBannerAd loaded.');
-          bannerCompleter.complete(ad as AdManagerBannerAd);
-        },
-        onAdFailedToLoad: (Ad ad, LoadAdError error) {
-          ad.dispose();
-          print('$AdManagerBannerAd failedToLoad: $error');
-          bannerCompleter.completeError(error);
-        },
-        onAdOpened: (Ad ad) => print('$AdManagerBannerAd onAdOpened.'),
-        onAdClosed: (Ad ad) => print('$AdManagerBannerAd onAdClosed.'),
-      ),
-    );
-    Future<void>.delayed(Duration(seconds: 1), () => _bannerAd.load());
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _bannerAd.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<AdManagerBannerAd>(
-      future: bannerCompleter.future,
-      builder:
-          (BuildContext context, AsyncSnapshot<AdManagerBannerAd> snapshot) {
-        Widget child;
-
-        switch (snapshot.connectionState) {
-          case ConnectionState.none:
-          case ConnectionState.waiting:
-          case ConnectionState.active:
-            child = Container();
-            break;
-          case ConnectionState.done:
-            if (snapshot.hasData) {
-              child = AdWidget(ad: _bannerAd);
-            } else {
-              child = Text('Error loading $AdManagerBannerAd');
-            }
-        }
-
-        return Container(
-          width: _bannerAd.sizes[0].width.toDouble(),
-          height: _bannerAd.sizes[0].height.toDouble(),
-          color: Colors.blueGrey,
-          child: child,
-        );
-      },
-    );
-  }
-}
-
-class NativeAdWidget extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() => NativeAdState();
-}
-
-class NativeAdState extends State<NativeAdWidget> {
-  late NativeAd _nativeAd;
-  final Completer<NativeAd> nativeAdCompleter = Completer<NativeAd>();
-
-  @override
-  void initState() {
-    super.initState();
-    _nativeAd = NativeAd(
-      adUnitId: NativeAd.testAdUnitId,
-      request: AdRequest(),
-      factoryId: 'adFactoryExample',
-      listener: NativeAdListener(
-        onAdLoaded: (Ad ad) {
-          print('$NativeAd loaded.');
-          nativeAdCompleter.complete(ad as NativeAd);
-        },
-        onAdFailedToLoad: (Ad ad, LoadAdError error) {
-          ad.dispose();
-          print('$NativeAd failedToLoad: $error');
-          nativeAdCompleter.completeError(error);
-        },
-        onAdOpened: (Ad ad) => print('$NativeAd onAdOpened.'),
-        onAdClosed: (Ad ad) => print('$NativeAd onAdClosed.'),
-      ),
-    );
-    Future<void>.delayed(Duration(seconds: 1), () => _nativeAd.load());
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _nativeAd.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<NativeAd>(
-      future: nativeAdCompleter.future,
-      builder: (BuildContext context, AsyncSnapshot<NativeAd> snapshot) {
-        Widget child;
-
-        switch (snapshot.connectionState) {
-          case ConnectionState.none:
-          case ConnectionState.waiting:
-          case ConnectionState.active:
-            child = Container();
-            break;
-          case ConnectionState.done:
-            if (snapshot.hasData) {
-              child = AdWidget(ad: _nativeAd);
-            } else {
-              child = Text('Error loading $NativeAd');
-            }
-        }
-
-        return Container(
-          width: 250,
-          height: 350,
-          color: Colors.blueGrey,
-          child: child,
-        );
-      },
-    );
-  }
-}
-
-void main() {
-  runApp(MyApp());
 }