+++ packages/google_mobile_ads/lib/src/ad_instance_manager.dart
@@ -501,6 +501,7 @@ class AdInstanceManager {
   }
 }
 
+@visibleForTesting
 class AdMessageCodec extends StandardMessageCodec {
   const AdMessageCodec();
 
@@ -590,14 +591,13 @@ class AdMessageCodec extends StandardMessageCodec {
         final String orientationStr =
             readValueOfType(buffer.getUint8(), buffer);
         final num width = readValueOfType(buffer.getUint8(), buffer);
-        final num height = readValueOfType(buffer.getUint8(), buffer);
         return AnchoredAdaptiveBannerAdSize(
           Orientation.values.firstWhere(
             (Orientation orientation) =>
                 describeEnum(orientation) == orientationStr,
           ),
           width: width.truncate(),
-          height: height.truncate(),
+          height: -1, // Unused value
         );
       case _valueSmartBannerAdSize:
         final String orientationStr =
@@ -711,7 +711,6 @@ class AdMessageCodec extends StandardMessageCodec {
       buffer.putUint8(_valueAnchoredAdaptiveBannerAdSize);
       writeValue(buffer, describeEnum(value.orientation));
       writeValue(buffer, value.width);
-      writeValue(buffer, value.height);
     } else if (value is SmartBannerAdSize) {
       buffer.putUint8(_valueSmartBannerAdSize);
       if (defaultTargetPlatform == TargetPlatform.iOS) {